bug(fix-eslint-errors):
add prop-types validation for different components
fix all eslint errors in CreateRecipe component
fix all eslint errors in CreateRecipeValidator class
[Finishes #154595824]
