+++ .eslintrc.json
@@ -30,6 +30,8 @@
             "components": [ "a", "Link" ],
             "aspects": ["invalidHref", "preferButton" ]
       }],
+      "jsx-a11y/click-events-have-key-events": 0,
+      "jsx-a11y/no-static-element-interactions": 0,
       "require-jsdoc": ["error", {
           "require": {
               "FunctionDeclaration": true,
