+++ client/src/js/screens/pages/CreateRecipe/validation/createRecipeValidator.js
@@ -1,6 +1,6 @@
 /**
  * Validates the data for creating a recipe
- * 
+ *
  * @export
  * @class CreateRecipeValidator
  */
@@ -23,8 +23,8 @@ export default class CreateRecipeValidator {
 
   /**
    * Check if the validation is valid
-   * 
-   * @returns 
+   *
+   * @returns {null} null
    * @memberof CreateRecipeValidator
    */
   isValid() {
@@ -35,8 +35,8 @@ export default class CreateRecipeValidator {
     this.validateProcedure();
     this.validateImage();
 
-    if (this.errors.title.length > 0 || 
-      this.errors.description.length > 0 || 
+    if (this.errors.title.length > 0 ||
+      this.errors.description.length > 0 ||
       this.errors.timeToCook.length > 0 ||
       this.errors.ingredients.length > 0 ||
       this.errors.procedure.length > 0 ||
@@ -50,16 +50,16 @@ export default class CreateRecipeValidator {
 
   /**
    * Validate the title creation
-   * 
+   * @returns {null} null
    * @memberof CreateRecipeValidator
    */
   validateTitle() {
     if (this.recipe.title) {
       if (this.recipe.title.length < 5) {
-        this.errors['title'].push('The title must be longer than 5 characters.');
+        this.errors.title.push('The title must be longer than 5 characters.');
       }
     } else {
-      this.errors['title'].push('The title is required.');
+      this.errors.title.push('The title is required.');
     }
   }
 
@@ -71,13 +71,13 @@ export default class CreateRecipeValidator {
   validateDescription() {
     if (this.recipe.description) {
       if (this.recipe.description.length < 5) {
-        this.errors['description'].push('The description must be longer than 5 characters.');
+        this.errors.description.push('The description must be longer than 5 characters.');
       }
       if (this.recipe.description.length > 220) {
-        this.errors['description'].push('The description must not be longer than 220 characters.');
+        this.errors.description.push('The description must not be longer than 220 characters.');
       }
     } else {
-      this.errors['description'].push('The description is required.');
+      this.errors.description.push('The description is required.');
     }
   }
 
@@ -95,30 +95,29 @@ export default class CreateRecipeValidator {
     }
   }
 
-  
   /**
    * Validate the ingredients field
    * @returns {null} no return
    */
   validateIngredients() {
-    let { ingredients } = this.recipe;
+    const { ingredients } = this.recipe;
 
-    ingredients.forEach(ingredient => {
+    ingredients.forEach((ingredient) => {
       if (ingredient.length < 3) {
-        this.errors['ingredients'].push('Oops ! Please make sure you actually typed in all ingredients.');
+        this.errors.ingredients.push('Oops ! Please make sure you actually typed in all ingredients.');
       }
     });
   }
-   /**
+  /**
    * Validate the ingredients field
    * @returns {null} no return
    */
   validateProcedure() {
-    let { procedure } = this.recipe;
+    const { procedure } = this.recipe;
 
     procedure.forEach((step) => {
       if (step.length < 3) {
-        this.errors['procedure'].push('Oops ! Please make sure you actually typed in all procedure steps.');
+        this.errors.procedure.push('Oops ! Please make sure you actually typed in all procedure steps.');
       }
     });
   }
